# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API MOBILE",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User information to login",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "vu.nguyentat@gmail.com"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "123456"
                                }
                            }
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "schema": {
                            "properties": {
                                "access_token": {
                                    "type": "string",
                                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                                },
                                "token_type": {
                                    "type": "string",
                                    "example": "bearer"
                                },
                                "expires_in": {
                                    "type": "integer",
                                    "example": 3600
                                }
                            },
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid email or password",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Username or password is incorrect"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "properties": {
                                "error": {
                                    "properties": {
                                        "email": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The email field is required.",
                                                "The email must be a valid email address."
                                            ]
                                        },
                                        "password": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The password field is required.",
                                                "The password must be at least 6 characters."
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Something went wrong."
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User information to register.",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Nguy\u1ec5n V\u0103n A"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "nguyenvana@gmail.com"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "0789876543"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully register",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "John Doe"
                                },
                                "username": {
                                    "type": "string",
                                    "example": "johndoe"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "john.doe@example.com"
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "0123456789"
                                },
                                "image": {
                                    "type": "string",
                                    "example": "Link image"
                                },
                                "address": {
                                    "type": "string",
                                    "example": "Ho Chi Minh"
                                },
                                "reward_point": {
                                    "type": "integer",
                                    "example": 100
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "properties": {
                                "error": {
                                    "properties": {
                                        "name": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The name field is required.",
                                                "The name may not be greater than 100 characters."
                                            ]
                                        },
                                        "email": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The email field is required.",
                                                "The email must be a valid email address.",
                                                "The email has already been taken."
                                            ]
                                        },
                                        "phone": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "The phone field is required.",
                                                "The phone has already been taken."
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Something went wrong."
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout the user and invalidate the token",
                "responses": {
                    "200": {
                        "description": "Successful logout",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Successfully logged out."
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User not authenticated",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Unauthenticated"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Something went wrong."
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "bearer_token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}